feat(tui): 按客户端的样子把视频、语音与文件画成卡片
附件消息此前只画 lark-cli 渲染出的文本，屏幕上就是 `<video key=… />` 这类
标签：既看不出是什么，也没有可操作的东西。

卡片改读 Feishu 发来的原始 body 而不是渲染结果——封面的 image_key 和 duration
只存在于 body 里，渲染文本没有。视频画成封面帧加 🎬 时长，语音只画 🎤 时长
（客户端就是把录音画成一条时长），其余文件画图标、文件名与大小。时长按播放器
的方式向上取整，否则 25.05s 的片子会显示成比客户端少一秒。

卡片整体就是点击目标：`o` 或鼠标点击打开已下载的那个文件。`o` 因此先问这条
消息画出了什么可点的东西（selectedZone），通话仍然是加入、两者共用同一条路径；
clickZone 随之带上 note，状态栏说明打开的是什么，而不是复述一串链接。没下载完
的附件不给目标，点了也没有文件可开。

会话列表摘要与引用行同样改走 attachGist：markup 里的 key 对人没有意义。片子的
文件名是一串序列号，比"[视频]"说得还少，所以只写类型不写名字。

humanBytes 转 int64 并补上 B 档：附件大小来自 resources.size_bytes，几百字节
的文件显示成 0.2 KB 没有意义。dataPath 在 dataDir 为空时不返回相对路径，否则
打开器会拿 TUI 恰好所在的工作目录去解释它。

+++ tui/yank.go
@@ -87,14 +87,14 @@ func renderIDs(src []yankSource) (string, string) {
 func renderRaw(src []yankSource) (string, string) {
 	if len(src) == 1 {
 		text := indentJSON(src[0].raw, "")
-		return text, "copied raw json · " + humanBytes(len(text))
+		return text, "copied raw json · " + humanBytes(int64(len(text)))
 	}
 	parts := make([]string, len(src))
 	for i, s := range src {
 		parts[i] = indentJSON(s.raw, "  ")
 	}
 	text := "[\n  " + strings.Join(parts, ",\n  ") + "\n]"
-	return text, fmt.Sprintf("copied %d json · %s", len(src), humanBytes(len(text)))
+	return text, fmt.Sprintf("copied %d json · %s", len(src), humanBytes(int64(len(text))))
 }
 
 // indentJSON makes a stored payload readable without a trip through jq. One
@@ -133,9 +133,9 @@ func renderContent(src []yankSource) (string, string) {
 	text := strings.Join(parts, "\n\n")
 	if len(src) == 1 {
 		if unrendered {
-			return text, "copied raw body (unrendered) · " + humanBytes(len(text))
+			return text, "copied raw body (unrendered) · " + humanBytes(int64(len(text)))
 		}
-		return text, "copied content · " + humanBytes(len(text))
+		return text, "copied content · " + humanBytes(int64(len(text)))
 	}
-	return text, fmt.Sprintf("copied %s · %s", plural(len(parts), "msg", "msgs"), humanBytes(len(text)))
+	return text, fmt.Sprintf("copied %s · %s", plural(len(parts), "msg", "msgs"), humanBytes(int64(len(text))))
 }
